+++ packages/core/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./components";
 
+export * from "./components/accordion/accordion-groups";
 export * from "./components/modal/modal.controller";