+++ packages/dso-toolkit/src/components/accordion/accordion.models.ts
@@ -7,6 +7,7 @@ export type AccordionSectionState = "success" | "info" | "warning" | "danger" | 
 export type AccordionHeading = "h2" | "h3" | "h4" | "h5";
 
 export interface Accordion<TemplateFnReturnType> {
+  group?: string;
   variant?: AccordionVariant;
   reverseAlign?: boolean;
   allowMultipleOpen?: boolean;