+++ storybook/src/components/accordion/accordion.core-template.ts
@@ -1,5 +1,5 @@
 import { Accordion } from "dso-toolkit";
-import { html, TemplateResult } from "lit-html";
+import { html, nothing, TemplateResult } from "lit-html";
 import { ifDefined } from "lit-html/directives/if-defined.js";
 import { ComponentImplementation } from "../../templates";
 
@@ -8,6 +8,7 @@ export const coreAccordion: ComponentImplementation<Accordion<TemplateResult>> =
   implementation: "core",
   template: () =>
     function accordionTemplate({
+      group,
       variant,
       reverseAlign,
       allowMultipleOpen,
@@ -17,6 +18,7 @@ export const coreAccordion: ComponentImplementation<Accordion<TemplateResult>> =
     }) {
       return html`
         <dso-accordion
+          group=${ifDefined(group)}
           variant=${ifDefined(variant)}
           ?reverse-align=${reverseAlign}
           ?allow-multiple-open=${allowMultipleOpen}
@@ -38,6 +40,33 @@ export const coreAccordion: ComponentImplementation<Accordion<TemplateResult>> =
             </dso-accordion-section>`
           )}
         </dso-accordion>
+        ${group
+          ? html`
+              <dso-accordion
+                group=${ifDefined(group)}
+                variant=${ifDefined(variant)}
+                ?reverse-align=${reverseAlign}
+                ?allow-multiple-open=${allowMultipleOpen}
+                @dsoToggleSection=${dsoToggleSection}
+                @dsoToggleSectionAnimationEnd=${dsoToggleSectionAnimationEnd}
+              >
+                ${sections.map(
+                  (section) => html`<dso-accordion-section
+                    ?open=${ifDefined(section.open)}
+                    handle-title=${section.handleTitle}
+                    heading=${section.heading}
+                    handle-url=${ifDefined(section.handleUrl)}
+                    state=${ifDefined(section.state)}
+                    status=${ifDefined(section.status)}
+                    icon=${ifDefined(section.icon)}
+                    attachment-count=${ifDefined(section.attachmentCount)}
+                  >
+                    ${section.content}
+                  </dso-accordion-section>`
+                )}
+              </dso-accordion>
+            `
+          : nothing}
       `;
     },
 };